+++ src/com/splunk/modinput/ModularInput.java
@@ -38,14 +38,12 @@ public abstract class ModularInput {
 	private static Map<String, String> transports = new HashMap<String, String>();
 	static {
 
-		transports.put("stdout",
-				"com.splunk.modinput.transport.STDOUTTransport");
+		transports.put("stdout", "com.splunk.modinput.transport.STDOUTTransport");
 		transports.put("hec", "com.splunk.modinput.transport.HECTransport");
 
 	}
 
-	protected Transport getTransportInstance(List<Param> params,
-			String stanzaName) {
+	protected Transport getTransportInstance(List<Param> params, String stanzaName) {
 		Transport instance = null;
 		String key = "stdout"; // default
 		HECTransportConfig hec = new HECTransportConfig();
@@ -78,8 +76,7 @@ public abstract class ModularInput {
 
 			} else if (param.getName().equals("hec_https")) {
 				try {
-					hec.setHttps(Boolean.parseBoolean(param.getValue().equals(
-							"1") ? "true" : "false"));
+					hec.setHttps(Boolean.parseBoolean(param.getValue().equals("1") ? "true" : "false"));
 				} catch (Exception e) {
 					logger.error("Can't determine hec https value, will revert to default value.");
 				}
@@ -95,8 +92,7 @@ public abstract class ModularInput {
 
 			} else if (param.getName().equals("hec_batch_mode")) {
 				try {
-					hec.setBatchMode(Boolean.parseBoolean(param.getValue()
-							.equals("1") ? "true" : "false"));
+					hec.setBatchMode(Boolean.parseBoolean(param.getValue().equals("1") ? "true" : "false"));
 				} catch (Exception e) {
 					logger.error("Can't determine batch_mode value, will revert to default value.");
 				}
@@ -115,19 +111,17 @@ public abstract class ModularInput {
 					logger.error("Can't determine max_batch_size_events value, will revert to default value.");
 				}
 
-			} else if (param.getName().equals(
-					"hec_max_inactive_time_before_batch_flush")) {
+			} else if (param.getName().equals("hec_max_inactive_time_before_batch_flush")) {
 				try {
-					hec.setMaxInactiveTimeBeforeBatchFlush(Long.parseLong(param
-							.getValue()));
+					hec.setMaxInactiveTimeBeforeBatchFlush(Long.parseLong(param.getValue()));
 				} catch (NumberFormatException e) {
-					logger.error("Can't determine max_inactive_time_before_batch_flush value, will revert to default value.");
+					logger.error(
+							"Can't determine max_inactive_time_before_batch_flush value, will revert to default value.");
 				}
 			}
 		}
 		try {
-			instance = (Transport) Class.forName(transports.get(key))
-					.newInstance();
+			instance = (Transport) Class.forName(transports.get(key)).newInstance();
 			instance.setStanzaName(stanzaName);
 			if (key.equalsIgnoreCase("hec"))
 				instance.init(hec);
@@ -143,8 +137,7 @@ public abstract class ModularInput {
 		try {
 			JAXBContext context = JAXBContext.newInstance(obj.getClass());
 			Marshaller marshaller = context.createMarshaller();
-			marshaller.setProperty(Marshaller.JAXB_FORMATTED_OUTPUT,
-					Boolean.TRUE);
+			marshaller.setProperty(Marshaller.JAXB_FORMATTED_OUTPUT, Boolean.TRUE);
 			marshaller.setProperty(Marshaller.JAXB_FRAGMENT, Boolean.TRUE);
 			StringWriter sw = new StringWriter();
 			marshaller.marshal(obj, sw);
@@ -163,8 +156,7 @@ public abstract class ModularInput {
 		try {
 			JAXBContext context = JAXBContext.newInstance(obj.getClass());
 			Marshaller marshaller = context.createMarshaller();
-			marshaller.setProperty(Marshaller.JAXB_FORMATTED_OUTPUT,
-					Boolean.TRUE);
+			marshaller.setProperty(Marshaller.JAXB_FORMATTED_OUTPUT, Boolean.TRUE);
 			marshaller.setProperty(Marshaller.JAXB_FRAGMENT, Boolean.TRUE);
 			StringWriter sw = new StringWriter();
 			marshaller.marshal(obj, sw);
@@ -228,19 +220,40 @@ public abstract class ModularInput {
 	protected boolean activationKeyCheck(String activationKey, String toHash) {
 
 		try {
-			
-			byte[] activationBytes = DatatypeConverter.parseHexBinary(activationKey);
 
-			MessageDigest md = MessageDigest.getInstance("MD5");
+			activationKey = activationKey.trim();
 
-			byte[] thedigest = md.digest(toHash.getBytes());
-
-			if (!MessageDigest.isEqual(thedigest, activationBytes)) {
-				logger.error("Activation key check failed");
-				return false;
+			if (activationKey.length() > 32) {
+				String activationHash = activationKey.substring(0, 32);
+				String activationTime = barbraStreisand(activationKey.substring(32));
+				long currentTime = System.currentTimeMillis()/1000;
+				MessageDigest md = MessageDigest.getInstance("MD5");
+				byte[] thedigest = md.digest((toHash + activationTime).getBytes());
+				byte[] activationBytes = DatatypeConverter.parseHexBinary(activationHash);
+				if (!MessageDigest.isEqual(thedigest, activationBytes)) {
+					logger.error("Activation key check failed.Please ensure that you copy/pasted the key correctly.");
+					return false;
+				} else if ((currentTime - Long.parseLong(activationTime)) > 604800) {
+					logger.error(
+							"Trial Activation key has now expired. Please visit http://www.baboonbones.com/#activation to purchase a non expiring key.");
+					return false;
+				} else {
+					logger.info("Activation key check passed");
+					return true;
+				}
 			} else {
-				logger.info("Activation key check passed");
-				return true;
+				MessageDigest md = MessageDigest.getInstance("MD5");
+				byte[] activationBytes = DatatypeConverter.parseHexBinary(activationKey);
+
+				byte[] thedigest = md.digest(toHash.getBytes());
+
+				if (!MessageDigest.isEqual(thedigest, activationBytes)) {
+					logger.error("Activation key check failed.Please ensure that you copy/pasted the key correctly.");
+					return false;
+				} else {
+					logger.info("Activation key check passed");
+					return true;
+				}
 			}
 		} catch (Throwable t) {
 			logger.error("Error performing activation key check : " + t.getMessage());
@@ -248,11 +261,14 @@ public abstract class ModularInput {
 		}
 
 	}
-	
+
+	protected String barbraStreisand(String original) {
+		return new StringBuilder(original).reverse().toString();
+	}
+
 	protected void init(String[] args) {
 
-		String loggingLevel = System.getProperty("splunk.logging.level",
-				"ERROR");
+		String loggingLevel = System.getProperty("splunk.logging.level", "ERROR");
 		logger.setLevel(Level.toLevel(loggingLevel));
 		logger.info("Initialising Modular Input");
 		try {
@@ -279,16 +295,14 @@ public abstract class ModularInput {
 				System.exit(2);
 			}
 		} catch (Exception e) {
-			logger.error("Error executing modular input : " + e.getMessage()
-					+ " : " + ModularInput.getStackTrace(e));
+			logger.error("Error executing modular input : " + e.getMessage() + " : " + ModularInput.getStackTrace(e));
 
 			System.exit(2);
 		}
 
 	}
 
-	protected void createTCPInput(Input input, int tcpPort, String index,
-			String sourcetype, String source) {
+	protected void createTCPInput(Input input, int tcpPort, String index, String sourcetype, String source) {
 
 		String host = input.getServer_host();
 		String uri = input.getServer_uri();
@@ -312,13 +326,12 @@ public abstract class ModularInput {
 			args.add("sourcetype", sourcetype);
 			args.add("source", sourcetype);
 
-			service.getInputs().create(String.valueOf(tcpPort), InputKind.Tcp,
-					args);
+			service.getInputs().create(String.valueOf(tcpPort), InputKind.Tcp, args);
 		}
 	}
 
-	protected String createHECInput(Input input, int hecPort, String index,
-			String sourcetype, String source, boolean https) {
+	protected String createHECInput(Input input, int hecPort, String index, String sourcetype, String source,
+			boolean https) {
 
 		String host = input.getServer_host();
 		String uri = input.getServer_uri();
@@ -352,10 +365,7 @@ public abstract class ModularInput {
 			for (Param param : params) {
 				if (param.getName().equals("disabled")) {
 					String val = param.getValue();
-					setDisabled(
-							stanza.getName(),
-							val.equals("0") || val.equalsIgnoreCase("false") ? false
-									: true);
+					setDisabled(stanza.getName(), val.equals("0") || val.equalsIgnoreCase("false") ? false : true);
 
 				}
 			}
@@ -400,17 +410,14 @@ public abstract class ModularInput {
 					try {
 						int index = stanza.indexOf("://");
 						// REST call to get state of input
-						com.splunk.Input input = service.getInputs().get(
-								stanza.substring(index + 3));
+						com.splunk.Input input = service.getInputs().get(stanza.substring(index + 3));
 						boolean isDisabled = input.isDisabled();
 						// update state map
 						setDisabled(stanza, isDisabled);
 						enabledCount += isDisabled ? 0 : 1;
 					} catch (Exception e) {
-						logger.error("Can't connect to Splunk REST API with the token ["
-								+ service.getToken()
-								+ "], either the token is invalid or SplunkD has exited : "
-								+ e.getMessage());
+						logger.error("Can't connect to Splunk REST API with the token [" + service.getToken()
+								+ "], either the token is invalid or SplunkD has exited : " + e.getMessage());
 					}
 				}
 				try {
@@ -433,8 +440,7 @@ public abstract class ModularInput {
 
 	}
 
-	protected synchronized static void setDisabled(String stanza,
-			boolean isDisabled) {
+	protected synchronized static void setDisabled(String stanza, boolean isDisabled) {
 
 		inputStates.put(stanza, isDisabled);
 
@@ -469,10 +475,9 @@ public abstract class ModularInput {
 						connectedToSplunk = true;
 						failCount = 0;
 					} catch (Exception e) {
-						logger.error("Probing socket connection to SplunkD failed.Either SplunkD has exited ,or if not,  check that your DNS configuration is resolving your system's hostname ("
-								+ this.splunkHost
-								+ ") correctly : "
-								+ e.getMessage());
+						logger.error(
+								"Probing socket connection to SplunkD failed.Either SplunkD has exited ,or if not,  check that your DNS configuration is resolving your system's hostname ("
+										+ this.splunkHost + ") correctly : " + e.getMessage());
 						failCount++;
 						connectedToSplunk = false;
 					} finally {
@@ -524,8 +529,7 @@ public abstract class ModularInput {
 				}
 			}
 		} catch (Throwable e) {
-			logger.error("Error setting JVM system propertys from string : "
-					+ propsString + " : " + getStackTrace(e));
+			logger.error("Error setting JVM system propertys from string : " + propsString + " : " + getStackTrace(e));
 		}
 
 	}